Commented servlet functionality and added API utilities RecommendationRetriever.java
