+++ src/main/java/com/google/impactdashboard/api_utilities/RecommendationRetriever.java
@@ -1,0 +1,11 @@
+package com.google.impactdashboard.api_utilities;
+
+import com.google.api.services.recommender.v1.Recommender;
+import com.google.api.services.recommender.v1.Recommender.Projects.Locations.Recommenders.Recommendations;
+
+/**
+ * Class that will call the Recommender API to get the full information for recommendations
+ */
+public class RecommendationRetriever {
+  // TODO: find out how to initialize recommender to send ListRecommendationsrequest
+}
