--- src/main/java/com/google/sps/servlets/GetProjectDataServlet.java
@@ -1,16 +1,0 @@
-package com.google.sps.servlets;
-
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@WebServlet("/get-project-data")
-public class GetProjectDataServlet extends HttpServlet {
-
-    @Override
-    public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-
-    }
-}
