--- src/main/java/com/google/sps/servlets/UpdateData.java
@@ -1,11 +1,0 @@
-package com.google.sps.servlets;
-
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@WebServlet("/update-data")
-public class UpdateData {
-}
