+++ src/test/java/com/google/impactdashboard/TestCI.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.sps;
+package com.google.impactdashboard;
 
 import static org.junit.Assert.assertEquals;
